+++ tests/Neo.L2.Abstractions.UnitTests/UT_NeoHubDeployReport.cs
@@ -309,8 +309,8 @@ public class UT_NeoHubDeployReport
             Assert.AreEqual(
                 "GatewayHostComposition.HasPendingPublication / PendingPublicationEpoch / "
                 + "AggregatorPendingCount / HasDurableOutbox / IsPublicationConfigured / "
-                + "IsEnabled / MaxAutomaticRetries / "
-                + "OutboxStatus / Aggregator / ReceiveBatch / "
+                + "IsEnabled / MaxAutomaticRetries / ProofSystem / ExpectedNetwork / "
+                + "HasL1RpcEndpoint / OutboxStatus / Aggregator / ReceiveBatch / "
                 + "PullAggregate (fails closed with durable outbox) / PublishAggregateAsync / "
                 + "RecoverPoisonedPublication / GetOperatorStatus / WriteOperatorStatusAsync / "
                 + "Metrics / CaptureMetricsSnapshot / ExportPrometheusMetrics / "